nomic draft
This is a draft of the nomic review. The nBTC review needs to be drafted tomorrow.

In the nBTC review we need to cover the bitcoin light client, circuit breaker mechanism, cryptoeconomic security.

for Nomic review, we need need to confirm if there is slashing logic for signatories signing a malicious pegout.